examples/permit-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
permit-sdk==2.0.2
python-dotenv==1.0.0
